Set up native EGL render loop for Oculus, Google and SVR flavors
