+++ app/src/oculusvr/AndroidManifest.xml
@@ -3,6 +3,7 @@
     <application>
         <meta-data android:name="com.samsung.android.vr.application.mode" android:value="vr_only"/>
         <activity android:name=".VRBrowserActivity" android:screenOrientation="landscape">
+            <meta-data android:name="android.app.lib_name" android:value="native-lib" />
         </activity>
     </application>
 </manifest>
